+++ tests/Spreads.Collections.Tests/MoveNextAsyncTests.cs
@@ -18,8 +18,7 @@ namespace Spreads.Collections.Tests {
         [Test]
         public void UpdateEventIsTriggered() {
             var sm = new SortedMap<DateTime, double>();
-            (sm as IUpdateable<DateTime, double>).OnData += (s,x) =>
-            {
+            (sm as IUpdateable<DateTime, double>).OnData += (s, x) => {
                 Console.WriteLine("Added {0} : {1}", x.Key, x.Value);
             };
 
@@ -29,10 +28,9 @@ namespace Spreads.Collections.Tests {
 
 
         [Test]
-        public void CouldMoveAsyncOnEmptySM()
-        {
+        public void CouldMoveAsyncOnEmptySM() {
             var sm = new SortedMap<DateTime, double>();
-            
+
             var c = sm.GetCursor();
 
             var moveTask = c.MoveNext(CancellationToken.None);
@@ -47,6 +45,9 @@ namespace Spreads.Collections.Tests {
 
         [Test]
         public void CouldReadSortedMapNewValuesWhileTheyAreAddedUsingCursor() {
+            var count = 1000000;
+            var sw = new Stopwatch();
+            sw.Start();
 
             var sm = new SortedMap<DateTime, double>();
             //sm.IsSynchronized = true;
@@ -56,7 +57,7 @@ namespace Spreads.Collections.Tests {
 
             var addTask = Task.Run(async () => {
                 await Task.Delay(50);
-                for (int i = 5; i < 10; i++) {
+                for (int i = 5; i < count; i++) {
                     sm.Add(DateTime.UtcNow.Date.AddSeconds(i), i);
                     //await Task.Delay(1);
                 }
@@ -69,12 +70,12 @@ namespace Spreads.Collections.Tests {
                     sum += c.CurrentValue;
                 }
                 Assert.AreEqual(10, sum);
-
+                var stop = DateTime.UtcNow.Date.AddSeconds(count - 1);
                 //await Task.Delay(50);
                 while (await c.MoveNext(CancellationToken.None)) {
                     sum += c.CurrentValue;
                     //Console.WriteLine("Current key: {0}", c.CurrentKey);
-                    if (c.CurrentKey == DateTime.UtcNow.Date.AddSeconds(10 - 1)) {
+                    if (c.CurrentKey == stop) {
                         break;
                     }
                 }
@@ -83,12 +84,65 @@ namespace Spreads.Collections.Tests {
 
             sumTask.Wait();
             addTask.Wait();
+            sw.Stop();
 
             double expectedSum = 0.0;
-            for (int i = 0; i < 10; i++) {
+            for (int i = 0; i < count; i++) {
                 expectedSum += i;
             }
             Assert.AreEqual(expectedSum, sum);
+
+            Console.WriteLine("Elapsed msec: {0}", sw.ElapsedMilliseconds - 50);
+        }
+
+
+        [Test]
+        public void CouldReadSortedMapNewValuesWhileTheyAreAddedUsingCursor_StartEmpty() {
+            var count = 100000;
+            var sw = new Stopwatch();
+            sw.Start();
+
+            var sm = new SortedMap<DateTime, double>();
+            //sm.IsSynchronized = true;
+            
+            var addTask = Task.Run(async () => {
+                await Task.Delay(50);
+                for (int i = 0; i < count; i++) {
+                    sm.Add(DateTime.UtcNow.Date.AddSeconds(i), i);
+                    //await Task.Delay(1);
+                }
+            });
+
+            double sum = 0.0;
+            var sumTask = Task.Run(async () => {
+                var c = sm.GetCursor();
+                while (c.MoveNext()) {
+                    sum += c.CurrentValue;
+                }
+                Assert.AreEqual(0, sum);
+                var stop = DateTime.UtcNow.Date.AddSeconds(count - 1);
+                //await Task.Delay(50);
+                while (await c.MoveNext(CancellationToken.None)) {
+                    sum += c.CurrentValue;
+                    //Console.WriteLine("Current key: {0}", c.CurrentKey);
+                    if (c.CurrentKey == stop) {
+                        break;
+                    }
+                }
+            });
+
+
+            sumTask.Wait();
+            addTask.Wait();
+            sw.Stop();
+
+            double expectedSum = 0.0;
+            for (int i = 0; i < count; i++) {
+                expectedSum += i;
+            }
+            Assert.AreEqual(expectedSum, sum);
+
+            Console.WriteLine("Elapsed msec: {0}", sw.ElapsedMilliseconds - 50);
         }
 
     }