+++ tests/Spreads.Collections.Tests/Profiler.cs
@@ -1,0 +1,16 @@
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Text;
+//using System.Threading.Tasks;
+
+//namespace Spreads.Collections.Tests {
+//    public static class Profiler {
+//        public static void Main()
+//        {
+//            var test = new MoveNextAsyncTests();
+//            test.CouldReadSortedMapNewValuesWhileTheyAreAddedUsingCursor();
+//            Console.ReadLine();
+//        }
+//    }
+//}